Fix: Make pipeline running but still another bug must be fixed

+++ src/data/image_postprocessing.py
@@ -238,7 +238,7 @@ class SuperpixelsProcessor:
                 # Get the most often repeated value
                 most_frequent_class_in_tile_segment = most_frequent_value_index.numpy()
                 # Create a tensor of ones with the shape of indices
-                ones = tf.ones((tf.shape(indices)[0],), dtype=tf.int64)
+                ones = tf.ones((tf.shape(indices)[0],), dtype=tf.uint8)
                 # Multiply the ones tensor by max_value
                 updates = ones * most_frequent_class_in_tile_segment
                 # Update the not_decoded_prediction tensor
@@ -258,7 +258,8 @@ class SuperpixelsProcessor:
             tensor contains integer values representing the segment labels (superpixel
             indices) assigned to each pixel in the image.
         """
-
+        # print(f"self.raw_image = {self.raw_image}")
+        # print(f"type = {type(self.raw_image)}")
         segments = slic(
             self.raw_image,
             **self.params_of_superpixels_postprocessing,
